fix: create multiple handlers
