+++ lambda/instructions/message.hbs
@@ -1,5 +1,5 @@
-target language: {{target_lng}}
-source language: {{source_lng}}
+target language: {{targetLanguage}}
+source language: {{sourceLanguage}}
 
 phrase - [Generate 1 intermediate phrase or 2-3 related sentences in source language from these categories:
 {{#if (eq weekday "Monday")}}
