+++ lambda/instructions/poll.hbs
@@ -1,5 +1,5 @@
-target language: {{target_lng}}
-source language: {{source_lng}}
+target language: {{targetLanguage}}
+source language: {{sourceLanguage}}
 
 Generate a language quiz question based on these criteria:
