+++ lambda/lib/config.js
@@ -1,8 +1,39 @@
 import path from "path";
 import fs from "fs";
 import Handlebars from "handlebars";
+import { fileURLToPath } from "url";
 
-export const registerHelpers = () => {
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
+let bridgeConfig = null;
+
+export function setupBridge() {
+  if (bridgeConfig) return bridgeConfig;
+
+  const BRIDGE = process.env.BRIDGE;
+  const [targetLanguage, sourceLanguage] = BRIDGE.split("-");
+  const BRIDGES_DIR = path.join(__dirname, "..", "bridges", BRIDGE);
+
+  registerHelpers();
+
+  bridgeConfig = {
+    targetLanguage,
+    sourceLanguage,
+    ...loadHBS(BRIDGES_DIR),
+  };
+
+  return bridgeConfig;
+}
+
+export function getBridgeTemplates() {
+  if (!bridgeConfig) {
+    setupBridge();
+  }
+  return bridgeConfig;
+}
+
+const registerHelpers = () => {
   Handlebars.registerHelper("pronunciationList", function (items) {
     if (!Array.isArray(items)) return "";
     return items.map((d) => `• "${d.word}" — ${d.explanation}`).join("\n");
@@ -20,7 +51,7 @@ export const registerHelpers = () => {
   });
 };
 
-export const loadHBS = (dir) => {
+const loadHBS = (dir) => {
   const messageInstructions = fs.readFileSync(
     path.join("instructions", "message.hbs"),
     "utf8",
